+++ app/build.gradle
@@ -4,6 +4,7 @@ plugins {
     id 'kotlin-kapt'
     id 'androidx.navigation.safeargs.kotlin'
     id 'com.google.gms.google-services'
+    id 'dagger.hilt.android.plugin'
 }
 
 android {
@@ -50,6 +51,11 @@ android {
         compose true
         dataBinding true
     }
+
+    // Allow references to generated code
+    kapt {
+        correctErrorTypes = true
+    }
 }
 
 dependencies {
@@ -58,10 +64,8 @@ dependencies {
     implementation 'com.google.android.material:material:1.6.1'
     implementation 'androidx.constraintlayout:constraintlayout:2.1.4'
 
-    implementation 'com.google.dagger:dagger:2.40.5'
     implementation 'androidx.legacy:legacy-support-v4:1.0.0'
     implementation 'androidx.lifecycle:lifecycle-livedata-ktx:2.5.1'
-    kapt 'com.google.dagger:dagger-compiler:2.40'
 
     implementation 'org.jetbrains.kotlinx:kotlinx-coroutines-android:1.6.1'
 
@@ -112,4 +116,9 @@ dependencies {
     implementation "com.google.accompanist:accompanist-appcompat-theme:0.25.1"
     // Constraints
     implementation "androidx.constraintlayout:constraintlayout-compose:1.0.1"
+
+    implementation("com.google.dagger:hilt-android:2.38.1")
+    kapt("com.google.dagger:hilt-android-compiler:2.38.1")
+    // for hiltViewModel
+    implementation("androidx.hilt:hilt-navigation-compose:1.0.0")
 }
