+++ app/src/main/java/com/example/ttanslateapp/domain/use_case/lists/AddNewListUseCase.kt
@@ -1,5 +1,6 @@
 package com.example.ttanslateapp.domain.use_case.lists
 
+import android.app.Application
 import com.example.ttanslateapp.domain.model.lists.ListItem
 import javax.inject.Inject
