+++ app/src/main/java/com/example/ttanslateapp/presentation/TranslateApp.kt
@@ -4,15 +4,11 @@ import android.app.Application
 import android.content.Context
 import androidx.appcompat.app.AppCompatDelegate
 import com.example.ttanslateapp.BuildConfig
-import com.example.ttanslateapp.di.DaggerApplicationComponent
-import dagger.hilt.android.AndroidEntryPoint
 import dagger.hilt.android.HiltAndroidApp
 import timber.log.Timber
 
 @HiltAndroidApp
 class TranslateApp : Application() {
-    val component = DaggerApplicationComponent.factory().create(this)
-
     init {
         instance = this
         AppCompatDelegate.setDefaultNightMode(AppCompatDelegate.MODE_NIGHT_NO);
