--- app/src/main/java/com/example/ttanslateapp/presentation/ViewModelFactory.kt
@@ -1,14 +1,0 @@
-package com.example.ttanslateapp.presentation
-
-import androidx.lifecycle.ViewModel
-import androidx.lifecycle.ViewModelProvider
-import javax.inject.Inject
-import javax.inject.Provider
-
-class ViewModelFactory @Inject constructor(
-    private val viewModelProvider: @JvmSuppressWildcards Map<Class<out ViewModel>, Provider<ViewModel>>
-):ViewModelProvider.Factory {
-    override fun <T : ViewModel> create(modelClass: Class<T>): T {
-        return viewModelProvider[modelClass]?.get() as T
-    }
-}
