+++ build.gradle
@@ -8,6 +8,7 @@ buildscript {
         def nav_version = "2.4.2"
         classpath "androidx.navigation:navigation-safe-args-gradle-plugin:$nav_version"
         classpath 'com.google.gms:google-services:4.3.13'
+        classpath "com.google.dagger:hilt-android-gradle-plugin:2.38.1"
     }
 }
